outsourcing routing
