+++ angular/practice-project/src/app/app.module.ts
@@ -26,23 +26,9 @@ import {ServerComponent} from "./servers/server/server.component";
 import {EditServerComponent} from "./servers/edit-server/edit-server.component";
 import {UsersComponent} from "./users/users.component";
 import {UserComponent} from "./users/user/user.component";
-import {RouterModule, Routes} from "@angular/router";
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 
-const appRoutes: Routes = [
-  {path: '', component: HomeComponent},
-  {path: 'users', component: UsersComponent},
-  {path: 'users/:id/:name', component: UserComponent},
-  {
-    path: 'servers', component: ServersComponent, children: [
-      {path: ':id/edit', component: EditServerComponent},
-      {path: ':id', component: ServerComponent}
-    ]
-  },
-  {path: 'server', redirectTo: '/servers', pathMatch: 'full'},
-  {path: 'not-found', component: PageNotFoundComponent},
-  {path: '**', redirectTo: '/not-found'}    // should be at the last route in the list
-]
+
 @NgModule({
   declarations: [
     AppComponent,
@@ -74,7 +60,7 @@ const appRoutes: Routes = [
     BrowserModule,
     AppRoutingModule,
     FormsModule,
-    RouterModule.forRoot(appRoutes)
+    AppRoutingModule
   ],
   providers: [],
   bootstrap: [AppComponent]
